# others/sdotodocservice-poc/src/main/webapp/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API for Document Service operations",
    "version" : "v1",
    "title" : "SDOTO Document Service API"
  },
  "host" : "127.0.0.1:8088",
  "basePath" : "/sdoto-docservice-war/rest",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/documentService/getDocument/{docId}" : {
      "get" : {
        "tags" : [ "documentService" ],
        "summary" : "Useful when you need to get document data when you know the document id",
        "description" : "The document id will never change",
        "operationId" : "getDocumentById",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "docId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/BookingDocumentDTO"
            }
          }
        }
      }
    },
...
    "/repository/getPdf/{docId}" : {
      "get" : {
        "tags" : [ "documentService" ],
        "summary" : "The getPdf returns a PDF document by its reference id",
        "description" : "You can know id using the DocumentServiceResource",
        "operationId" : "getPdf",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/pdf" ],
        "parameters" : [ {
          "name" : "docId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/InputStream"
            }
          }
        }
      }
    }
  },
  "definitions" : {
...
    "BookingDocumentDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "state" : {
          "type" : "string",
          "enum" : [ "CREATED", "SENT", "OPENED", "CLOSED", "FAIL" ]
        },
        "parking" : {
          "type" : "string"
        },
        "creationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "openDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "closeDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "sentDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "docReference" : {
          "type" : "string"
        }
      }
    },
...
    "InputStream" : {
      "type" : "object"
    },
...
  }
}